+++ CHANGELOG.md
@@ -11,10 +11,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Change font to a monospace font
 * Wrap news to allow multi line 
 * Added a darker color for news signatures ([#4](https://github.com/gastbob40/InfinityNews/issues/4))
+* Added new way to refresh newsgroups by swiping down ([#5](https://github.com/gastbob40/InfinityNews/issues/5))
 
 ### Fix
 
 * Upgrade auto unwrap method, by checking next chars (as  `>`, `\n`, `[` etc) ([#3](https://github.com/gastbob40/InfinityNews/issues/3))
+
+### Removed
+
+* Removed the refresh button in the popover ([#5](https://github.com/gastbob40/InfinityNews/issues/5))
 
 ## [0.1.1] - 2020-09-26